+++ pkgs/collection/CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.19.1-wip
+
 - Add `IterableMapEntryExtension` for working on `Map` as a list of pairs, using
   `Map.entries`.
+- Optimize equality and hash code for maps by using `update` and a `values`
+  iterator to avoid extra lookups.
 
 ## 1.19.1